BinanceFeedScraper/requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-restframework==0.0.1
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
lxml==4.9.3
outcome==1.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
urllib3==2.0.4
wsproto==1.2.0
